♻️ Refactor: fetch 로직을 coreFetch와 publicFetch로 분리
